program to make record of students and allow to access record of students.
This is a Mini project to make a record of 20 students whose name,roll number,sessional marks and semester marks are input then he/shecan  see thier records by first varifying the roll number.If roll  is not correct then a goto statment display a option to enter roll number again.
A students  can access thier
1.personal details
2.sessional marks (total & percentage)
3.semster marks (total &percentage)
4.lis of topper of first year.
5.Exit option to exit from program

This project is done by moving cursor on the screen.