program to understand static data member and member function bt making  a simple program
Static:-static variabe or class variable doesn't take object memory,it takes different memory location and liftime entire the program .It is autimatically initialized to 0.
static fuction is used to access static private member.

In this program we make qurbani class and static data member to count number of member.